+++ packages/experimental/Button/src/FAB/FAB.ios.tsx
@@ -1,3 +1,3 @@
-import { FAB } from './FABMobile';
+import { FAB } from './FABCore';
 
 export default FAB;
